# openapi.yaml
openapi: 3.0.3
info:
  title: Liquor Distribution OS API
  version: 1.0.0
  description: |
    REST API for the Liquor Distribution Operating System.
    Covers customers, orders, invoices, collections, deliveries, inventory, and sync.
    All mutating endpoints require an `Idempotency-Key` header.
servers:
  - url: https://api.example.com/v1
    description: Production
  - url: https://staging.api.example.com/v1
    description: Staging
security:
  - bearerAuth: []
...
paths:
...
  /products:
    get:
      tags: [Catalog]
      parameters:
        - { in: query, name: q, schema: { type: string } }
        - { in: query, name: category, schema: { type: string } }
        - { in: query, name: active, schema: { type: boolean, default: true } }
      responses: { '200': { description: ok } }
  /products/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    get:
      tags: [Catalog]
      responses: { '200': { description: ok } }
...
  /visits/{id}/payment:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    post:
      tags: [Collections]
      summary: Record payment with allocations
      parameters:
        - { in: header, name: Idempotency-Key, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PaymentInput' }
      responses:
        '201':
          description: created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Payment' }
...
  /payments/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    get:
      tags: [Collections]
      responses: { '200': { description: ok } }
    patch:
      tags: [Collections]
      summary: Edit payment (admin only; creates audit entry)
      responses: { '200': { description: ok } }
  /payments/{id}/verify:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    post:
      tags: [Collections]
      summary: Accounts verifies (cheque cleared / bank credited)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [result]
              properties:
                result: { type: string, enum: [verified, bounced] }
                note: { type: string }
      responses: { '200': { description: ok } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PaymentInput:
      type: object
      required: [amount, mode, allocations, gps]
      properties:
        amount: { type: number }
        mode: { type: string, enum: [cash, cheque, bank, upi] }
        mode_ref: { type: string }
        cheque_date: { type: string, format: date }
        proof_image_url: { type: string, format: uri }
        gps: { $ref: '#/components/schemas/GeoPoint' }
        allocations:
          type: array
          items:
            type: object
            required: [invoice_id, amount]
            properties:
              invoice_id: { type: string }
              amount: { type: number }
    Payment:
      type: object
      properties:
        id: { type: string }
        amount: { type: number }
        mode: { type: string }
        mode_ref: { type: string }
        verification_status: { type: string, enum: [pending, verified, bounced] }
        collected_at: { type: string, format: date-time }
        locked_at: { type: string, format: date-time, nullable: true }
...
    GeoPoint:
      type: object
      required: [lat, lng]
      properties:
        lat: { type: number }
        lng: { type: number }
        accuracy_m: { type: number }
...